rename componentPages to componentViews
