+++ src/componentViews/examples/BlogPosts.jsx
