+++ src/componentViews/examples/CheckoutPage.jsx
