+++ src/componentViews/examples/Ecommerce.jsx
