+++ src/componentViews/examples/Pricing.jsx
